+++ AD/ADmulti.py
